src/main/webapp/app/entities/all4qmsMsMetaind/meta-recurso/delete/meta-recurso-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { provideHttpClient, HttpResponse } from '@angular/common/http';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { MetaRecursoService } from '../service/meta-recurso.service';

import { MetaRecursoDeleteDialogComponent } from './meta-recurso-delete-dialog.component';

describe('MetaRecurso Management Delete Component', () => {
  let comp: MetaRecursoDeleteDialogComponent;
  let fixture: ComponentFixture<MetaRecursoDeleteDialogComponent>;
  let service: MetaRecursoService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [MetaRecursoDeleteDialogComponent],
      providers: [provideHttpClient(), NgbActiveModal],
    })
      .overrideTemplate(MetaRecursoDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(MetaRecursoDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(MetaRecursoService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
